add new story!
We are working on a group project for a full stack bootcamp and we are using your API. This is just a test to see if this works so...if it does, we will be adding a number of stories to your template. We are trying to use your template so please let me know if I've done something that doesn't work.

Thanks! 
-Shawn Flanigan
feel free to reach out to me at shawnmflanigan@gmail.com