I am adding a new story to your templates.json
Hello, and a happy New Years to you! 

I am working with Shawn Flanigan on the same project and am going to be adding between 5 & 10 stories to this templates.JSON, but wanted to run it by you as well as confirm that the formatting is correct. I just added the "A Spooky Campfire Story" story and am pretty sure that I followed the pre-existing formatting guidelines, but please let me know of any changes that are needed. Thank you for your time.